requirements.txt
# Requisitos principales
PyQt5>=5.15.2
PyMongo>=4.0.0
pymongo[srv]  # Para conexiones MongoDB con URI de conexión

# Temas y estilos
QDarkStyle>=3.0.3

# Herramientas para empaquetado
PyInstaller>=4.5.0

# Utilidades
python-dotenv>=0.19.0